refactor: rename create-draft-publish-helper.service to draft-publish-helper.service and update imports
