+++ src/index.ts
@@ -335,7 +335,7 @@ export * from './services/action-metadata.service'
 export * from './services/authentication.service'
 export * from './services/bcrypt.service'
 export * from './services/computed-fields/uuid-external-id-computed-field-provider.service'
-export * from './services/create-draft-publish-helper.service'
+export * from './services/draft-publish-helper.service'
 export * from './services/crud-helper.service'
 export * from './services/crud.service'
 export * from './services/email-template.service'
